# generated/routes/profiles.ts
// Auto-generated by RouteGenerator. Do not edit manually.
import { Hono } from 'hono';
import type { AppEnv } from 'schematic-pg/api/types';
import { validateJson, validateParam, validateQuery } from 'schematic-pg/api/middleware/validate';
import { notFoundResponse } from 'schematic-pg/api/middleware/errors';
import { buildReadQuery } from 'schematic-pg/api/utils/read-query';
import { parseIncludeQuery } from 'schematic-pg/api/utils/include-query';
import { omitFields } from 'schematic-pg/api/utils/omit-fields';
import { shapeResponse, shapeResponseMany } from 'schematic-pg/api/utils/response-shape';
import {
  ProfileCreateSchema,
  ProfileUpdateSchema,
  ProfileParamSchema,
  ProfileListQuerySchema,
  ProfileGetQuerySchema,
  PROFILE_LIST_QUERY_FIELDS,
  PROFILE_SORTABLE_FIELDS,
  PROFILE_INCLUDABLE_RELATIONS,
  PROFILE_OMIT_FIELDS,
  API_OMIT_FIELDS_BY_MODEL,
  API_RELATION_TARGETS,
} from '../schemas/validation.js';

const router = new Hono<AppEnv>();

router.get('/', validateQuery(ProfileListQuerySchema), async (c) => {
  const db = c.get('db');
  const query = c.req.valid('query');
  const { where, orderBy, take, skip, include } = buildReadQuery(
    query,
    PROFILE_LIST_QUERY_FIELDS,
    PROFILE_SORTABLE_FIELDS,
    PROFILE_INCLUDABLE_RELATIONS,
  );
  const rows = await db.profile.findMany({
    where,
    orderBy,
    take,
    skip,
    include,
  });
  return c.json(shapeResponseMany(rows, 'Profile', API_OMIT_FIELDS_BY_MODEL, API_RELATION_TARGETS));
});

router.get('/:id', validateParam(ProfileParamSchema), validateQuery(ProfileGetQuerySchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const query = c.req.valid('query');
  const include = query.include
    ? parseIncludeQuery(query.include, PROFILE_INCLUDABLE_RELATIONS)
    : undefined;
  const row = await db.profile.findUnique({ id: params.id }, { include });
  if (!row) {
    return notFoundResponse(c);
  }
  return c.json(shapeResponse(row, 'Profile', API_OMIT_FIELDS_BY_MODEL, API_RELATION_TARGETS));
});

router.post('/', validateJson(ProfileCreateSchema), async (c) => {
  const db = c.get('db');
  const body = c.req.valid('json');
  const row = await db.profile.create(body);
  return c.json(omitFields(row, PROFILE_OMIT_FIELDS), 201);
});

router.put('/:id', validateParam(ProfileParamSchema), validateJson(ProfileUpdateSchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const body = c.req.valid('json');
  const row = await db.profile.update({ where: { id: params.id }, data: body });
  return c.json(omitFields(row, PROFILE_OMIT_FIELDS));
});

router.delete('/:id', validateParam(ProfileParamSchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const row = await db.profile.delete({ id: params.id });
  return c.json(omitFields(row, PROFILE_OMIT_FIELDS));
});

export default router;